feat: Add custom hook to trigger submit.
